# Implement support for optimizing away the dynamic registrar.