--- a/runtime/Delegates.cs.t4
+++ b/runtime/Delegates.cs.t4
@@ -68,9 +68,23 @@ namespace ObjCRuntime {
 
 <# } #>
 
+		[BindingImpl (BindingImplOptions.Optimizable)] // To inline the Runtime.DynamicRegistrationSupported code if possible.
 		static void RegisterDelegates (InitializationOptions* options)
 		{
-<# foreach (var d in delegates) { #>
+<# foreach (var d in delegates) {
+	if (d.OnlyDynamicUsage) continue; #>
+			options->Delegates-><#= d.SimpleEntryPoint #> = GetFunctionPointer (new <#= d.SimpleEntryPoint #>_delegate (<#= d.SimpleEntryPoint #>));
+<# } #>
+
+			// The linker will remove this condition (and the subsequent method call) if possible
+			if (DynamicRegistrationSupported)
+				RegisterDelegatesDynamic (options);
+		}
+
+		static void RegisterDelegatesDynamic (InitializationOptions* options)
+		{
+<# foreach (var d in delegates) {
+	if (!d.OnlyDynamicUsage) continue; #>
 			options->Delegates-><#= d.SimpleEntryPoint #> = GetFunctionPointer (new <#= d.SimpleEntryPoint #>_delegate (<#= d.SimpleEntryPoint #>));
 <# } #>
 		}