--- a/runtime/runtime-generated.h.t4
+++ b/runtime/runtime-generated.h.t4
@@ -22,6 +22,8 @@ extern "C" {
 #endif
 
 <# foreach (var d in delegates) {
+		if (d.OnlyDynamicUsage)
+			continue;
 #>
 <#= d.CReturnType #>
 <#= d.EntryPoint #> (<#= d.CArgumentSignature #>);