--- a/src/Foundation/NSObject2.cs
+++ b/src/Foundation/NSObject2.cs
@@ -282,6 +282,7 @@ namespace Foundation {
 
 		[Export ("conformsToProtocol:")]
 		[Preserve ()]
+		[BindingImpl (BindingImplOptions.Optimizable)]
 		public virtual bool ConformsToProtocol (IntPtr protocol)
 		{
 			bool does;