--- a/src/ObjCRuntime/Blocks.cs
+++ b/src/ObjCRuntime/Blocks.cs
@@ -85,6 +85,9 @@ namespace ObjCRuntime {
 
 		void SetupBlock (Delegate trampoline, Delegate userDelegate, bool safe)
 		{
+			if (!Runtime.DynamicRegistrationSupported)
+				throw ErrorHelper.CreateError (8026, "BlockLiteral.SetupBlock is not supported when the dynamic registrar has been linked away.");
+
 			// We need to get the signature of the target method, so that we can compute
 			// the ObjC signature correctly (the generated method that's actually
 			// invoked by native code does not have enough type information to compute
@@ -275,7 +278,11 @@ namespace ObjCRuntime {
 			// with the proper reference count.
 			BlockLiteral block = new BlockLiteral ();
 			if (signature == null) {
-				block.SetupBlock ((Delegate) handlerDelegate, (Delegate) @delegate);
+				if (Runtime.DynamicRegistrationSupported) {
+					block.SetupBlock ((Delegate) handlerDelegate, (Delegate) @delegate);
+				} else {
+					throw ErrorHelper.CreateError (8026, $"BlockLiteral.GetBlockForDelegate with a null signature is not supported when the dynamic registrar has been linked away (delegate type: {@delegate.GetType ().FullName}).");
+				}
 			} else {
 				block.SetupBlockImpl ((Delegate) handlerDelegate, (Delegate) @delegate, true, signature);
 			}