--- a/src/ObjCRuntime/Class.cs
+++ b/src/ObjCRuntime/Class.cs
@@ -27,6 +27,7 @@ namespace ObjCRuntime {
 
 		internal IntPtr handle;
 
+		[BindingImpl (BindingImplOptions.Optimizable)]
 		internal unsafe static void Initialize (Runtime.InitializationOptions* options)
 		{
 			var map = options->RegistrationMap;
@@ -444,6 +445,7 @@ namespace ObjCRuntime {
 		/*
 		Type must have been previously registered.
 		*/
+		[BindingImpl (BindingImplOptions.Optimizable)] // To inline the Runtime.DynamicRegistrationSupported code if possible.
 #if !XAMCORE_2_0 && !MONOTOUCH // Accidently exposed this to public, can't break API
 		public
 #else