@rolfbjarne 2018-02-15
Wrench results: https://wrench.internalx.com/Wrench/ViewLane.aspx?lane_id=5746&host_id=392&revision_id=929081

All test failures are unrelated:

* linkall/Mac/Debug: [linkall/Mac/Debug build failure](https://github.com/xamarin/maccore/issues/642)
* Mono.Security/Mac/Modern: [Test failure on bots: Mono.Security/Mac/Modern: MonoTests.Mono.Security.Authenticode.AuthenticodeDeformatterTest.VerifySignedAssembly](https://github.com/xamarin/xamarin-macios/issues/3207)
* MSBuild/iOS: [[XI]WatchOS apps fail to build with error "Undefined symbols for architecture armv7k" using XI with mono-2017-12 support](https://github.com/mono/mono/issues/6948)
* MTouch tests: [MTouch tests: Xamarin.MTouch.RebuildTest_WithExtensions failure](https://github.com/xamarin/maccore/issues/641)
* xammac/Release: [[jenkins] XM/introspection random crashes](https://github.com/xamarin/maccore/issues/568)